# public/apidocs/api-docs.json
{  
  "swagger":"2.0",
  "info":{  
    "title":"Cents API",
    "description":"Examine your life decisions RESTfully.",
    "version":"2.0.0"
  },
  "paths":{  
    "/careers":{  
      "get":{  
        "tags":[  
          "careers"
        ],
        "responses":{  
          "200":{  
            "description":"Records were found."
          },
          "404":{  
            "description":"No records were found."
          }
        },
        "description":"",
        "summary":"Returns names of all available careers.",
        "operationId":"Api::V2::CareersController#show_two"
      }
    },
    "/careers/compare":{  
      "post":{  
        "tags":[  
          "careers"
        ],
        "responses":{  
          "200":{  
            "description":"Records were found."
          },
          "400":{  
            "description":"No careers were in the careers array."
          },
          "404":{  
            "description":"No data on some careers."
          }
        },
        "description":"This route is intended for internal use. The order parameter is optional, but if you use it, you must use it in exactly two career objects. Its value is either 1 or 2.",
        "summary":"Returns data on two careers.",
        "operationId":"Api::V2::CareersController#show_two",
        "parameters":[  
          {  
            "in":"body",
            "description":"object with array of career objects and optional operation",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/Career_body"
            }
          }
        ]
      }
    },
...
    "/cost_of_living/compare":{  
      "post":{  
        "tags":[  
          "cost of living"
        ],
        "responses":{  
          "200":{  
            "description":"Records were found."
          },
          "400":{  
            "description":"No objects were in the locations array."
          },
          "404":{  
            "description":"No data on city or state for some locations."
          }
        },
        "description":"This route is intended for internal use. The order parameter is optional, but if you use it, you must use it in exactly two location objects. Its value is either 1 or 2.",
        "summary":"Returns cost of living information for two locations.",
        "operationId":"Api::V2::Coli#show_two",
        "parameters":[  
          {  
            "in":"body",
            "description":"object with array of location objects and optional operation",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/Cost_of_living_body"
            }
          }
        ]
      }
    },
...
    "/degrees/compare":{  
      "post":{  
        "tags":[  
          "degrees"
        ],
        "responses":{  
          "200":{  
            "description":"Records were found."
          },
          "400":{  
            "description":"No objects were in the degrees array."
          },
          "404":{  
            "description":"No data for some degrees."
          }
        },
        "description":"This route is intended for internal use. The order parameter is optional, but if you use it, you must use it in exactly two degree objects. Its value is either 1 or 2.",
        "summary":"Returns degree information for two degrees.",
        "operationId":"Api::V2::DegreesController#show_two",
        "parameters":[  
          {  
            "in":"body",
            "description":"object with array of degree objects and optional operation",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/Degree_body"
            }
          }
        ]
      }
    },
    "/degrees/{level}/{name}":{  
      "get":{  
        "tags":[  
          "degrees"
        ],
        "responses":{  
          "200":{  
            "description":"Records were found."
          },
          "404":{  
            "description":"Degree does not exist."
          }
        },
        "description":"",
        "summary":"Returns information on all degrees that start with {level} and have the name {name}.",
        "operationId":"Api::V2::DegreesController#show_level_name",
        "parameters":[  
          {  
            "in":"path",
            "description":"a degree level",
            "name":"level",
            "required":true,
            "type":"string"
          },
          {  
            "in":"path",
            "description":"a degree name",
            "name":"name",
            "required":true,
            "type":"string"
          }
        ]
      }
    },
...
    "/schools/compare":{  
      "post":{  
        "tags":[  
          "schools"
        ],
        "responses":{  
          "200":{  
            "description":"Records were found."
          },
          "400":{  
            "description":"No schools were in the schools array."
          },
          "404":{  
            "description":"No data on some schools."
          }
        },
        "description":"This route is intended for internal use. The order parameter is optional, but if you use it, you must use it in exactly two school objects. Its value is either 1 or 2.",
        "summary":"Returns data on two schools.",
        "operationId":"Api::V2::SchoolsController#show_two",
        "parameters":[  
          {  
            "in":"body",
            "description":"object with array of school objects and optional operation",
            "name":"body",
            "required":true,
            "schema":{  
              "$ref":"#/definitions/School_body"
            }
          }
        ]
      }
    },
...
  },
  "host":"trycents.com",
  "basePath":"/api/v2",
  "schemes":[  
    "https"
  ],
  "definitions":{  
    "Career_body":{  
      "type":"object",
      "properties":{  
        "operation":{  
          "type":"string"
        },
        "careers":{  
          "type":"array",
          "items":{  
            "$ref":"#/definitions/Career"
          }
        }
      },
      "required":[  
        "careers"
      ]
    },
    "Career":{  
      "required":[  
        "name"
      ],
      "properties":{  
        "name":{  
          "type":"string"
        },
        "order":{  
          "type":"integer",
          "enum":[  
            1,
            2
          ]
        }
      }
    },
    "Cost_of_living_body":{  
      "type":"object",
      "properties":{  
        "operation":{  
          "type":"string"
        },
        "locations":{  
          "type":"array",
          "items":{  
            "$ref":"#/definitions/Location"
          }
        }
      },
      "required":[  
        "locations"
      ]
    },
    "Location":{  
      "required":[  
        "city",
        "state"
      ],
      "properties":{  
        "city":{  
          "type":"string"
        },
        "state":{  
          "type":"string"
        },
        "order":{  
          "type":"integer",
          "enum":[  
            "1",
            "2"
          ]
        }
      }
    },
    "Degree_body":{  
      "type":"object",
      "properties":{  
        "operation":{  
          "type":"string"
        },
        "degrees":{  
          "type":"array",
          "items":{  
            "$ref":"#/definitions/Degree"
          }
        }
      },
      "required":[  
        "degrees"
      ]
    },
    "Degree":{  
      "required":[  
        "name",
        "level"
      ],
      "properties":{  
        "name":{  
          "type":"string"
        },
        "level":{  
          "type":"string"
        },
        "order":{  
          "type":"integer",
          "enum":[  
            "1",
            "2"
          ]
        }
      }
    },
...
    "School_body":{  
      "type":"object",
      "properties":{  
        "operation":{  
          "type":"string"
        },
        "schools":{  
          "type":"array",
          "items":{  
            "$ref":"#/definitions/School"
          }
        }
      },
      "required":[  
        "schools"
      ]
    },
    "School":{  
      "required":[  
        "name"
      ],
      "properties":{  
        "name":{  
          "type":"string"
        },
        "order":{  
          "type":"integer",
          "enum":[  
            "1",
            "2"
          ]
        }
      }
    },
...
  }
}